+++ compiler-core/src/parse/tests.rs
@@ -1907,3 +1907,25 @@ fn operator_in_pattern_size() {
 fn correct_precedence_in_pattern_size() {
     assert_parse!("let assert <<size, payload:size(size + 2 * 8)>> = <<>>");
 }
+
+#[test]
+fn function_definition_angle_generics_error() {
+    assert_module_error!("fn id<T>(x: T) { x }");
+}
+
+#[test]
+fn type_angle_generics_usage_error() {
+    assert_error!("let list: List<Int> = []");
+}
+
+#[test]
+fn type_angle_generics_definition_error() {
+    assert_module_error!(
+        r#"
+type Maybe<A> {
+    Some(A)
+    None
+}
+"#
+    );
+}